Added folders for protocol, experiment and other types of files
git-svn-id: file:///home/svn/shoola-omero/trunk@6150 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/treeviewer/browser/BrowserUI.java
@@ -1140,8 +1140,8 @@ class BrowserUI
 				}
 			} else if (node instanceof TreeFileSet) {
 				if (((TreeFileSet) node).getType() == index) {
-					if (value instanceof Integer) 
-						node.setNumberItems((Integer) value);
+					if (value instanceof Long) 
+						node.setNumberItems((Long) value);
 				}
 			}
